requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.9.5
aiosignal==1.2.0
altair==5.5.0
annotated-types==0.6.0
anyio==4.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
Babel==2.11.0
beautifulsoup4==4.12.3
bleach==4.1.0
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.0.9
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
comm==0.2.1
dataclasses-json==0.6.5
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
jupyterlab-widgets==3.0.10
langchain==0.3.7
langchain-community==0.3.6
langchain-core==0.3.18
langchain-experimental==0.3.3
langchain-openai==0.2.8
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.143
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.16.0
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.3
numexpr==2.10.1
numpy==1.26.4
openai==1.54.0
orjson==3.10.7
overrides==7.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pip==24.3.1
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==5.29.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==18.1.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydantic-settings==2.6.1
pydeck==0.9.1
Pygments==2.15.1
PyMuPDF==1.25.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
qtconsole==5.6.0
QtPy==2.4.1
referencing==0.30.2
regex==2024.9.11
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.10.6
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
semantic-router==0.0.72
Send2Trash==1.8.2
setuptools==75.1.0
sip==6.7.12
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.34
stack-data==0.2.0
streamlit==1.40.2
streamlit-calendar==1.2.1
sympy==1.13.1
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.2.1
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.47.1
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
widgetsnbextension==4.0.10
win-inet-pton==1.1.0
yarl==1.11.0
